Разбил камни, все работает
